Extract process-control coordination into modules/process_control.py
Phase 6 of breaking up the quickstart.py monolith. Moves the
~1185-line process-control layer — maintenance window resolution,
pending-start queueing, kometa/imagemaid process discovery and
launch/stop, run-context tracking, and meta.log/sidecar marker
writing — into modules/process_control.py.

The routes that depend on it (start/stop-kometa, kometa-status,
tail-log, shutdown, restart, start/stop-imagemaid, imagemaid-status)
stay in quickstart.py: they're each individually monkeypatch-tested
via qs_module._foo, same pattern as the routes kept in place during
the kometa-install and imagemaid phases.

Several functions inside the new module are themselves individually
monkeypatched by tests as dependencies of other moved functions
(_maintenance_guard_loop, the marker-writing functions,
_launch_imagemaid_command). Moving the caller and callee into the same
module together breaks those patches, since the bare-name call now
resolves through process_control.py's globals instead of
quickstart.py's. Fixed with the same deferred `import quickstart`
bridge already used for _get_active_work_blocker — applied to
_get_maintenance_window_live/_from_db, _get_run_context,
_get_imagemaid_run_context, _peek/pop_pending_kometa_start,
_find_running_imagemaid_process, _launch_kometa_command,
_suspend/resume_process_tree, _write_quickstart_maintenance_marker,
_append_quickstart_meta_log_line, and _reset_imagemaid_runtime_env.

Running the full suite against this surfaced 20 pre-existing failures
that predate this phase entirely (confirmed against the unmodified
develop tip before this branch touched anything):

- apply_kometa_selection (modules/kometa_install.py, from the Kometa
  root/install phase) used current_app, a context-dependent proxy,
  where the original inline code referenced the literal Flask app
  object directly — broke 17 tests that exercise wizard-status code
  with no active Flask context. Fixed by guarding the app.config
  writes with has_app_context(), mirroring the existing
  has_request_context() guard around the session writes in the same
  function.
- _probe_kometa_root_state and _get_imagemaid_supported_options (from
  the Kometa-install and ImageMaid phases) were never re-exported into
  quickstart.py, even though tests monkeypatch them directly.
- Several marker-writing functions read app.config["VERSION_CHECK"]
  unconditionally via current_app, failing outside any Flask context;
  added a small _get_version_info() helper that degrades to {} instead
  of raising.

Net: quickstart.py shrinks by roughly 1230 lines.
